log app contents on plan failure
