+++ src/nixpacks/app.rs
@@ -35,7 +35,6 @@ impl App {
     }
 
     /// Check if a file exists
-
     pub fn includes_file(&self, name: &str) -> bool {
         self.source.join(name).is_file()
     }
@@ -94,7 +93,6 @@ impl App {
     }
 
     /// Check if a path matching a glob exists
-
     pub fn has_match(&self, pattern: &str) -> bool {
         match self.find_files(pattern) {
             Ok(v) => !v.is_empty(),
@@ -139,7 +137,6 @@ impl App {
     }
 
     /// Check if a directory exists
-
     pub fn includes_directory(&self, name: &str) -> bool {
         self.source.join(name).is_dir()
     }
@@ -195,7 +192,6 @@ impl App {
     }
 
     /// Get the path in the container to an asset defined in `static_assets`.
-
     pub fn asset_path(&self, name: &str) -> String {
         format!("{}{}", ASSETS_DIR, name)
     }
